refactor: divide interface and impl
refactor: divide interface and impl
